replaced : Model Class をapp/ から app/Infrastructure/Eloquent/ へ移動した
