+++ app/Infrastructure/Eloquent/Book.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App;
+namespace App\Infrastructure\Eloquent;
 
 use Illuminate\Database\Eloquent\Model;
 use Alsofronie\Uuid\Uuid32ModelTrait;
 
-class BookAnchors extends Model
+class Book extends Model
 {
     use Uuid32ModelTrait;
