+++ app/Infrastructure/Eloquent/BookAnchors.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App;
+namespace App\Infrastructure\Eloquent;
 
 use Illuminate\Database\Eloquent\Model;
 use Alsofronie\Uuid\Uuid32ModelTrait;
 
-class Anchor extends Model
+class BookAnchors extends Model
 {
     use Uuid32ModelTrait;
