+++ app/Infrastructure/Eloquent/Tag.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App;
+namespace App\Infrastructure\Eloquent;
 
 use Illuminate\Database\Eloquent\Model;
 use Alsofronie\Uuid\Uuid32ModelTrait;
 
-class Book extends Model
+class Tag extends Model
 {
     use Uuid32ModelTrait;
